Add themes handling as ActionPlugin
